creat creat_ui.py
把建設ui的程式碼分離開來
